Refactor default layout, extrude functions and components
